Move tests which require rabbitMQ to be running to integration_tests
