--- .ci/faf-rabbitmq.env
@@ -1,9 +1,0 @@
-RABBITMQ_PID_FILE=/var/lib/rabbitmq/pid
-
-RABBITMQ_DEFAULT_USER=admin
-RABBITMQ_DEFAULT_PASS=banana
-RABBITMQ_DEFAULT_VHOST=/
-
-RABBITMQ_LOBBYSERVER_USER=faf-lobby
-RABBITMQ_LOBBYSERVER_PASS=banana
-RABBITMQ_LOBBYSERVER_VHOST=/faf-lobby
