--- .ci/teardown-rabbitmq.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-docker stop faf-rabbitmq-test
