+++ tests/conftest.py
@@ -230,14 +230,8 @@ async def rating_service(database, player_service):
 
     await service.shutdown()
 
-@pytest.fixture(scope="session")
-def ensure_rabbitmq_is_running():
-    subprocess.call(".ci/init-rabbitmq.sh")
-    yield
-    subprocess.call(".ci/teardown-rabbitmq.sh")
-
 @pytest.fixture
-async def message_queue_service(ensure_rabbitmq_is_running):
+async def message_queue_service():
     service = MessageQueueService()
     await service.initialize()
