+++ tests/unit_tests/test_message_queue_service.py
@@ -11,40 +11,8 @@ from server.message_queue_service import MessageQueueService
 pytestmark = pytest.mark.asyncio
 
 
-@with_logger
-class Consumer:
-    def __init__(self):
-        self._callback = mock.Mock()
-
-    async def initialize(self):
-        self.connection = await aio_pika.connect(
-            "amqp://{user}:{password}@localhost/{vhost}".format(
-                user=config.MQ_USER, password=config.MQ_PASSWORD, vhost=config.MQ_VHOST
-            )
-        )
-        channel = await self.connection.channel()
-        exchange = await channel.declare_exchange(
-            "test_exchange", aio_pika.ExchangeType.TOPIC
-        )
-        self.queue = await channel.declare_queue("test_queue", exclusive=True)
-
-        await self.queue.bind(exchange, routing_key="#")
-        self.consumer_tag = await self.queue.consume(self.callback)
-
-    def callback(self, message):
-        self._logger.debug("Received message %r", message)
-        self._callback()
-
-    def callback_count(self):
-        return self._callback.call_count
-
-    async def shutdown(self):
-        await self.queue.cancel(self.consumer_tag)
-        await self.connection.close()
-
-
 @pytest.fixture
-async def mq_service(ensure_rabbitmq_is_running):
+async def mq_service():
     service = MessageQueueService()
     await service.initialize()
 
@@ -55,17 +23,7 @@ async def mq_service(ensure_rabbitmq_is_running):
     await service.shutdown()
 
 
-@pytest.fixture
-async def consumer(ensure_rabbitmq_is_running):
-    consumer = Consumer()
-    await consumer.initialize()
-
-    yield consumer
-
-    await consumer.shutdown()
-
-
-async def test_initialize(ensure_rabbitmq_is_running):
+async def test_initialize():
     service = MessageQueueService()
     await service.initialize()
     await service.shutdown()
@@ -80,26 +38,7 @@ async def test_publish(mq_service):
     await mq_service.publish(exchange_name, routing_key, payload, delivery_mode)
 
 
-async def test_publish_wrong_exchange(mq_service, ensure_rabbitmq_is_running):
-    bad_exchange = "nonexistent_exchange"
-    with pytest.raises(KeyError):
-        await mq_service.publish(bad_exchange, "", {})
-
-
-async def test_consumer_receives(mq_service, consumer):
-    payload = {"msg": "test message"}
-    exchange_name = "test_exchange"
-    routing_key = "test.routing.key"
-    delivery_mode = aio_pika.DeliveryMode.NOT_PERSISTENT
-
-    await mq_service.publish(exchange_name, routing_key, payload, delivery_mode)
-
-    await asyncio.sleep(0.1)
-
-    assert consumer.callback_count() == 1
-
-
-async def test_incorrect_credentials(mocker, ensure_rabbitmq_is_running):
+async def test_incorrect_credentials(mocker):
     mocker.patch("server.message_queue_service.config.MQ_PASSWORD", "bad_password")
     service = MessageQueueService()
     service._logger = mock.Mock()
@@ -120,7 +59,7 @@ async def test_incorrect_credentials(mocker, ensure_rabbitmq_is_running):
     await service.shutdown()
 
 
-async def test_incorrect_username(mocker, ensure_rabbitmq_is_running):
+async def test_incorrect_username(mocker):
     mocker.patch("server.message_queue_service.config.MQ_USER", "bad_user")
     service = MessageQueueService()
     service._logger = mock.Mock()
@@ -129,7 +68,7 @@ async def test_incorrect_username(mocker, ensure_rabbitmq_is_running):
     service._logger.warning.assert_called()
 
 
-async def test_incorrect_port(mocker, ensure_rabbitmq_is_running):
+async def test_incorrect_port(mocker):
     mocker.patch("server.message_queue_service.config.MQ_PORT", 1)
     service = MessageQueueService()
     service._logger = mock.Mock()
@@ -138,22 +77,10 @@ async def test_incorrect_port(mocker, ensure_rabbitmq_is_running):
     service._logger.warning.assert_called()
 
 
-async def test_incorrect_vhost(mocker, ensure_rabbitmq_is_running):
+async def test_incorrect_vhost(mocker):
     mocker.patch("server.message_queue_service.config.MQ_VHOST", "bad_vhost")
     service = MessageQueueService()
     service._logger = mock.Mock()
 
     await service.initialize()
     service._logger.warning.assert_called()
-
-
-async def test_reconnect(mq_service, ensure_rabbitmq_is_running):
-    await mq_service.declare_exchange("test_topic", aio_pika.ExchangeType.TOPIC)
-    await mq_service.declare_exchange("test_direct", aio_pika.ExchangeType.DIRECT)
-
-    await mq_service.reconnect()
-
-    assert "test_topic" in mq_service._exchanges
-    assert mq_service._exchange_types["test_topic"] == aio_pika.ExchangeType.TOPIC
-    assert "test_direct" in mq_service._exchanges
-    assert mq_service._exchange_types["test_direct"] == aio_pika.ExchangeType.DIRECT
